+++ src/test/java/org/mmarini/railways2/model/TrainRunningThroughSignalTest.java
@@ -31,7 +31,6 @@ package org.mmarini.railways2.model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mmarini.Tuple2;
-import org.mmarini.railways2.model.geometry.Direction;
 import org.mmarini.railways2.model.geometry.StationBuilder;
 import org.mmarini.railways2.model.geometry.StationMap;
 import org.mmarini.railways2.model.routes.Entry;